+++ __test__/main.test.ts
@@ -458,9 +458,3 @@ describe("importOnestopToBasecamp", () => {
         expect(deleteScheduleEntryMock).toHaveBeenCalledTimes(0);
     });
 });
-
-describe("loadDataFromOnestopIntoScriptProperties", () => {
-    it("should when", () => {
-
-    });
-});
